Merge sell buy games with login-logout-register

+++ src/main/java/id/ac/ui/cs/advprog/adproggameshop/utility/GameDTO.java
@@ -7,14 +7,14 @@ import lombok.Setter;
 public class GameDTO {
     private Long productId;
     private String name;
-    private int price;
+    private double price;
     private int quantity;
     private String category;
     private Long ownerUserId;
     private String ownerUsername;
 
     // Constructor
-    public GameDTO(Long productId, String name, int price,  int quantity, String category, Long ownerUserId, String ownerUsername) {
+    public GameDTO(Long productId, String name, double price,  int quantity, String category, Long ownerUserId, String ownerUsername) {
         this.productId = productId;
         this.name = name;
         this.price = price;
